load_book_done

+++ src/Book.java
@@ -3,13 +3,13 @@ public class Book {
     private int bookID;
     private String bookName;
     private String authorName;
-    private double quantity;
+    private int quantity;
     private String status;
 
     public Book(){
     }
 
-    public Book(int bookID, String bookName, String authorName, double quantity, String status) {
+    public Book(int bookID, String bookName, String authorName, int quantity, String status) {
         this.bookID = bookID;
         this.bookName = bookName;
         this.authorName = authorName;
@@ -41,11 +41,11 @@ public class Book {
         this.authorName = authorName;
     }
 
-    public double getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(double quantity) {
+    public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
